Modified Logic in the Course Service class

+++ src/main/java/org/example/studentmanagementsystem/service/CourseService.java
@@ -5,6 +5,8 @@ import org.example.studentmanagementsystem.entity.Course;
 import org.example.studentmanagementsystem.entity.Student;
 import org.example.studentmanagementsystem.repository.CourseRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -19,28 +21,34 @@ public class CourseService {
         this.courseRepository = courseRepository;
     }
 
-    public void save(Course course){
+    public ResponseEntity<Course> save(Course course){
         courseRepository.save(course);
+        return new ResponseEntity<>(course, HttpStatus.CREATED);
     }
 
-    public Course findById(int id){
+    public ResponseEntity<Course> findById(int id){
         Course course = courseRepository.findById(id);
-        if(course == null){
-            throw new NullPointerException("course of id " + id + " doesn't exist");
+        if (course == null) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
-        return course;
+        return new ResponseEntity<>(course, HttpStatus.OK);
 
     }
 
-    public void delete(int id){
+    public ResponseEntity<Void> delete(int id){
+        if(courseRepository.findById(id) == null){
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
         courseRepository.delete(id);
+        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
 
-    public List<Course> findAll(){
+    public ResponseEntity<List<Course>> findAll(){
         List<Course> courses = courseRepository.findAll();
-        if(courses == null)
-            throw new NullPointerException("List of courses is empty and currently NULL");
-        return courses;
+        if (courses.isEmpty()) {
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        }
+        return new ResponseEntity<>(courses, HttpStatus.OK);
     }
 
     public void update(Course course){
